feat: moving pagination and filter to hook
